fix: address code-review findings on the orderbook fix
Multi-agent review of the stale-book commits surfaced two real-money
regressions in the defense-in-depth code plus correctness/polish items.
All actionable findings fixed; the core float-accumulation fix was sound.

P1 (real-money paths):
- ws.py: reconnect never reset ws_owned, so a book stayed WS-owned across a
  dropped connection — if the replayed snapshot was delayed/lost, both REST
  repair paths were locked out (permanent staleness). Now release ownership in
  _replay_subscriptions before resubscribing; the fresh snapshot re-asserts it.
- Frontend REST seeds (EventView, MarketCard) clobbered the live exact-float WS
  book with rounded REST ints on card re-expand / slow reload — the frontend
  had no ws_owned counterpart. Now producer-form setQueryData: seed only when
  the cache is empty; the authoritative WS snapshot still overwrites.
- orders.py read raw float qty into the int-typed sanity-guard input on the
  buy/sell-now overpay path. Now reads int_levels().

P2:
- live_state/WebSocketProvider: presence now derived from the same round() the
  display uses (keep iff rounds to >= 1), so a stored level can never render or
  be selected as qty 0. Removes the 0.5-epsilon-vs-round boundary asymmetry.
- ws_manager: a snapshot now drops same-ticker deltas already queued in the
  flush window, so the browser can't apply a pre-snapshot delta onto the
  post-snapshot baseline.
- Cross-referenced the duplicated drop/round logic between backend and frontend.

P3:
- events.py seed guard keys on ws_owned (not both-sides-non-empty) and seeds
  concurrently via asyncio.gather; fixed quantity_at/delta-int stale docstrings
  and the orphaned broadcast_queue docstring.

Tests (the unanimous gap — the fix's core invariant was untested):
- Fractional delta accumulation (exact sum, no truncation, no drift over 1000
  deltas), presence-matches-display rounding, int_levels, ws_owned set/release,
  _poll_one drop-for-ws-owned, resync_locked skip-when-authoritative,
  refresh_now_await not entering the FAR schedule, fractional delta_fp parsing.

134 tests pass; mypy --strict and tsc clean.

+++ backend/src/kalshi/ws_wire.py
@@ -239,7 +239,7 @@ def parse_kalshi_ws_message(raw: dict) -> KalshiWsMessage | None:
 
     Conversion responsibilities:
       - orderbook_snapshot: yes/no arrays of [price_dollar_str, qty] → list[BookLevel]
-      - orderbook_delta:    price_dollars + delta_fp (string) → price_cents + delta (int)
+      - orderbook_delta:    price_dollars + delta_fp (string) → price_cents (int) + delta (float)
       - fill / user_order:  yes_price_dollars / yes_price → cents
     """
     msg_type = raw.get("type")
